sang 20-10-2025

+++ BackEnd/product_service/src/main/java/Myproject/product_service/entity/Product.java
@@ -25,7 +25,7 @@ public class Product {
 
     private String productBrand;
     private String productName;
-    private double productScreenSize;
+    private float productScreenSize;
     private String productColor;
     private int productRam;
     private int productRom;
